+++ aggregator/src/light_aggregator.rs
@@ -560,6 +560,7 @@ mod tests {
                 jubjub: true,
                 poseidon: true,
                 sha256: true,
+                sha512: false,
                 secp256k1: false,
                 bls12_381: false,
                 base64: false,