+++ circuits/examples/bitcoin_signature.rs
@@ -141,6 +141,7 @@ impl Relation for BitcoinSigExample {
             jubjub: false,
             poseidon: false,
             sha256: true,
+            sha512: false,
             secp256k1: true,
             bls12_381: false,
             base64: false,