+++ circuits/examples/ecc_ops.rs
@@ -71,6 +71,7 @@ impl Relation for EccExample {
             jubjub: true,
             poseidon: true,
             sha256: false,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,