+++ circuits/examples/hybrid_mt.rs
@@ -332,6 +332,7 @@ impl Relation for HybridMtExample {
             jubjub: false,
             poseidon: true,
             sha256: true,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,