+++ circuits/examples/json_verification.rs
@@ -135,6 +135,7 @@ impl Relation for AtalaJsonECDSA {
             jubjub: false,
             poseidon: false,
             sha256: true,
+            sha512: false,
             secp256k1: true,
             bls12_381: false,
             base64: true,