+++ circuits/examples/membership.rs
@@ -89,6 +89,7 @@ impl Relation for MembershipExample {
             jubjub: false,
             poseidon: true,
             sha256: false,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,