+++ circuits/examples/poseidon.rs
@@ -45,6 +45,7 @@ impl Relation for PoseidonExample {
             jubjub: false,
             poseidon: true,
             sha256: false,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,