+++ circuits/examples/rsa_signature.rs
@@ -79,6 +79,7 @@ impl Relation for RSASignatureCircuit {
             jubjub: false,
             poseidon: false,
             sha256: false,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,