+++ circuits/examples/schnorr_sig.rs
@@ -147,6 +147,7 @@ impl Relation for SchnorrExample {
             jubjub: true,
             poseidon: true,
             sha256: false,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,