+++ circuits/examples/sha_preimage.rs
@@ -54,6 +54,7 @@ impl Relation for ShaPreImageCircuit {
             jubjub: false,
             poseidon: false,
             sha256: true,
+            sha512: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,