+++ circuits/src/hash/mod.rs
@@ -15,3 +15,4 @@
 
 pub mod poseidon;
 pub mod sha256;
+pub mod sha512;